pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splatrix"
version = "0.1.0"
description = "Convert video files to 3D Gaussian Splats"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["gaussian-splatting", "3d-reconstruction", "nerfstudio", "video", "ply"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "PyQt6>=6.6.0",
    "PyQt6-Qt6>=6.6.0",
    "PyQt6-WebEngine>=6.6.0",
    "PyQt6-QmlImport>=6.6.0",
    "psutil>=5.9.0",
    "PyYAML>=6.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "plyfile>=1.0.0",
    "av>=12.0.0",
]

[project.optional-dependencies]
nerfstudio = [
    "nerfstudio>=1.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
dev = [
    "ruff",
    "pytest",
]

[project.scripts]
splatrix = "splatrix.main_qml:main"

[project.urls]
Homepage = "https://github.com/splatrix/splatrix"
Documentation = "https://splatrix.github.io/splatrix"
Repository = "https://github.com/splatrix/splatrix"
Issues = "https://github.com/splatrix/splatrix/issues"

[tool.setuptools.packages.find]
include = ["splatrix*"]

[tool.setuptools.package-data]
splatrix = [
    "qml/*.qml",
    "qml/qmldir",
    "qml/fonts/*.ttf",
    "qml/icons/*.svg",
    "viewer/*.html",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
